fix: fixed model deletion in settings ref: presentation moved to the `ui` folder
